// client/lootlion-web/src/app/api/generated/model/addMemberBody.ts
/**
 * Lootlion API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MemberRole } from './memberRole';


export interface AddMemberBody { 
    memberUserId?: string;
    role?: MemberRole;
}