// client/lootlion-web/src/app/api/generated/model/userSearchHitDto.ts
/**
 * Lootlion API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSearchHitDto { 
    userId?: string;
    displayName?: string | null;
    email?: string | null;
}